Allow range validators to accept 64-bit min/max values

+++ service/src/main/java/org/whispersystems/textsecuregcm/grpc/validators/Range.java
@@ -5,7 +5,7 @@
 
 package org.whispersystems.textsecuregcm.grpc.validators;
 
-public record Range(int min, int max) {
+public record Range(long min, long max) {
   public Range {
     if (min > max) {
       throw new IllegalArgumentException("invalid range values: expected min <= max but have [%d, %d],".formatted(min, max));
